fix: remove ef from application
