+++ src/Dfe.PlanTech.Application/Caching/Models/QueryCacher.cs
@@ -1,7 +1,4 @@
-using System.Security.Cryptography;
-using System.Text;
 using Dfe.PlanTech.Application.Caching.Interfaces;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Caching.Memory;
 
 namespace Dfe.PlanTech.Application.Caching.Models;
@@ -11,19 +8,12 @@ public class QueryCacher: IQueryCacher
     private MemoryCache _cache = new(new MemoryCacheOptions());
     private const int CacheDurationMinutes = 30;
 
-    private static string GetCacheKey(IQueryable query)
-    {
-        var queryString = query.ToQueryString();
-        var hash = SHA256.HashData(Encoding.UTF8.GetBytes(queryString));
-        return Convert.ToBase64String(hash);
-    }
-
     public async Task<TResult> GetOrCreateAsyncWithCache<T, TResult>(
+        string key,
         IQueryable<T> queryable,
         Func<IQueryable<T>, CancellationToken, Task<TResult>> queryFunc,
         CancellationToken cancellationToken = default)
     {
-        var key = GetCacheKey(queryable);
         return await _cache.GetOrCreateAsync(key, entry =>
         {
             entry.AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(CacheDurationMinutes);
@@ -31,17 +21,7 @@ public class QueryCacher: IQueryCacher
         }) ?? await queryFunc(queryable, cancellationToken);
     }
 
-    public async Task<List<T>> ToListAsyncWithCache<T>(IQueryable<T> queryable, CancellationToken cancellationToken = default)
-    {
-        return await GetOrCreateAsyncWithCache(queryable, (q, ctoken) => q.ToListAsync(ctoken), cancellationToken);
-    }
-
-    public async Task<T?> FirstOrDefaultAsyncWithCache<T>(IQueryable<T> queryable, CancellationToken cancellationToken = default)
-    {
-        return await GetOrCreateAsyncWithCache(queryable, (q, ctoken) => q.FirstOrDefaultAsync(ctoken), cancellationToken);
-    }
-
-    public void ClearCmsCache()
+    public void ClearCache()
     {
         _cache.Dispose();
         _cache = new MemoryCache(new MemoryCacheOptions());
